Port MORGUE to Windows: it was Clang-only, not macOS-only
bytebeat.h included <stdatomic.h> unconditionally, outside any __cplusplus
guard, and nine C++ files both included it and called the C atomics directly.
_Atomic in C++ is a Clang extension, so MSVC rejected it and g++ would have
too -- this never would have built on Linux either. That single header was the
root cause, not the operating system.

  bb_atomic.h     BB_ATOMIC(T) -> std::atomic<T> in C++, _Atomic(T) in C, with
                  static_asserts on size, alignment and lock-freedom so an ABI
                  mismatch fails at compile time instead of silently corrupting
                  cross-thread state. bytebeat.h's declarations now sit inside
                  one extern "C" region; six externs were outside it and would
                  have been C++-mangled in the GUI and C-linked in the engine.

  bb_platform.c   The POSIX calls engine.c made with no platform conditionals
                  at all: mkdir(p,0755), rename() as the atomic-replace step,
                  clock_gettime(CLOCK_MONOTONIC) inside bb_engine_render, and
                  narrow fopen(). rename() over an existing file FAILS on the
                  Windows CRT, so the session would have saved exactly once and
                  then silently never again -- Main.cpp ignored the return
                  value. Paths now widen through UTF-8 so a non-ASCII username
                  works.

  -fwrapv         MSVC has no equivalent and silently dropped it (D9002), so
                  the defined-overflow guarantee that the Makefile calls "the
                  instrument" had quietly evaporated on Windows. Fixed in the
                  code instead: the negative left-shifts in engine.c and dsp.c
                  now shift through the matching unsigned type. The suite
                  proves the output is bit-identical. The flag stays for
                  GCC/Clang as defence in depth, not as the guarantee.

  tests/          The 2,966-check suite lived inside main.c -- the file the TUI
                  retirement deletes -- called audio_self_test() through ALSA,
                  and hardcoded mkdtemp("/tmp/..."), which MSVC does not have.
                  Extracted whole to a JUCE-free target that configures and
                  runs with third_party/JUCE completely empty. Nothing dropped
                  to make it compile. 21 new checks cover the port itself:
                  paths with spaces and non-ASCII characters, and CRLF
                  round-trips.

  Session.h       Six call sites rebuilt "$HOME/MORGUE" by hand and a seventh
                  painted the literal "~/MORGUE/session.conf" into the title
                  bar -- a lie on Windows. One answer now, from the engine.

  Autosave        bb_config_save() was called from exactly one place: the
                  MainComponent destructor. No timer, no dirty flag. A crash
                  took the whole session, while the field manual promised
                  "the session autosaves". Now dirty-flagged and debounced,
                  with a save on focus loss, and the return value is checked.

  CMakeLists      juce_add_gui_app (bare add_executable linked
                  /SUBSYSTEM:CONSOLE while JUCE defines WinMain), /utf-8 for
                  the 71 non-ASCII lines in app/ that have no BOM, m linked
                  only on UNIX AND NOT APPLE, /experimental:c11atomics,
                  JUCE_WEB_BROWSER=0, a real default build type, and a clear
                  message instead of an opaque failure when the submodule is
                  empty.

  tools/exhume.py archive.org acquisition: advancedsearch.php (the scrape API
                  reports total=13 where the true numFound is 10374 for guests),
                  self-resolved storage nodes with failover (/download
                  round-robins onto CDN nodes that 500 on Range requests),
                  md5 verification against the Archive's own checksums, ffmpeg
                  transcode to 44.1k mono, and a provenance sidecar that
                  records who declared the licence.

Verified on Windows 11 / MSVC 19.44.35228: engine and suite compile clean,
all 2,987 checks pass, MORGUE.exe links, runs, opens a 48 kHz device and
renders its console.

+++ dsp.c
@@ -54,7 +54,14 @@ void post_reset(PostState *ps)
 
 int32_t dsp_dcblock(PostState *ps, int32_t x)
 {
-    int64_t y = (((int64_t)(x - ps->dc_x1)) << 16)
+    /* The scaling shift goes through uint64_t. `x - dc_x1` is negative for
+     * half of every waveform, and a signed left shift of a negative value is
+     * undefined behaviour -- Clang only behaved because the Makefile passes
+     * -fwrapv, and MSVC has no -fwrapv to pass. Shifting the unsigned bit
+     * pattern and reinterpreting it produces the identical value on every
+     * compiler; this is the same U()/S() trick expr.c plays with the VM's
+     * arithmetic. Nothing about the filter changes. */
+    int64_t y = (int64_t)((uint64_t)(int64_t)(x - ps->dc_x1) << 16)
               + ((ps->dc_y1 * DC_R) >> 16);
     ps->dc_x1 = x;
     ps->dc_y1 = y;
@@ -89,7 +96,10 @@ static inline int32_t saturate(int32_t x)
  * see the comment in expr.c for why the naive int32 version silently dies. */
 static inline int32_t onepole(int64_t *st, int32_t x, int32_t c)
 {
-    int64_t target = (int64_t)x << 8;
+    /* Unsigned shift for the same reason as dsp_dcblock: x is a signed
+     * sample, it is negative half the time, and `<<` on a negative signed
+     * value is undefined. Same bits, defined everywhere. */
+    int64_t target = (int64_t)((uint64_t)(int64_t)x << 8);
     *st += ((target - *st) * c) >> 8;
     return (int32_t)(*st >> 8);
 }
